% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_diag.R
\name{add_diag}
\alias{add_diag}
\title{Add to the diagonal of a square matrix.}
\usage{
add_diag(x, epsilon = sqrt(.Machine$double.eps), check = TRUE)
}
\arguments{
\item{x}{A square matrix, either of type double or integer.}

\item{epsilon}{Values to be added to the main diagonal of \code{x}, either a single numeric
value or a numeric vector. Its type must match the type of \code{x}.}

\item{check}{Should checks of the input be performed? Not doing so is faster.}
}
\value{
Noting, the matrix is modified in place.
}
\description{
Adds a constant or a vector to the diagonal of a numeric square matrix.
}
\details{
This function modifies the matrix \code{x} in place for getting maximum speed by
avoiding making copies.
}
\examples{
d <- diag(5L)
add_diag(d, 1)
diag(d)

# Speed comparison:
d <- diag(1000L)
epsilon <- 1e-6
microbenchmark::microbenchmark(times = 100L,
add_diag(d, epsilon),
add_diag(d, epsilon, check = FALSE),
diag(d) <- diag(d) + epsilon # canonical approach
)

}
\author{
Dominik Mueller
}
