% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_adjust.R
\name{age_adjust}
\alias{age_adjust}
\title{Calculate age-adjusted incidence rates}
\usage{
age_adjust(data, exposure_var, age_var, event_var, time_var)
}
\arguments{
\item{data}{A dataframe containing the variables used. Dataframe.}

\item{exposure_var}{Column in data that contains the exposure variable. Numeric or categorical vector.}

\item{age_var}{Column in data that contains the age variable in years. Numerical vector.}

\item{event_var}{Column in data that contains the event indicator. Binary numerical vector, 0 = no event, 1 = event.}

\item{time_var}{Column in data that contains the follow-up time indicator in days. Numerical vector.}
}
\value{
A dataframe with the following columns
\itemize{
\item \code{Treatment}:  Treatment groups
\item \code{IR}:  Age- and sex-adjusted incidence rate per 100 person year
\item \code{Lower_CI}:  Lower 95\% confidence interval
\item \code{Upper_CI}:  Upper 95\% confidence interval
}
}
\description{
This function uses a poisson model to adjust for age.
It is primarily intended for use when the variable of interest is sex.
It returns a dataframe with estimated incidence rate per 100 person-years.
Uses marginaleffects::avg_predictions to obtain marginal population standardized estimates
}
\examples{
colon_death <- colon[colon$etype == 2, ]
# The row above selects rows with event indicator and follow-up time for death
# Adjusting for age where the grouping variable is sex.
age_adjust(colon_death, sex, age, status, time)
age_adjust(colon[colon$etype == 2, ], sex, age, status, time)

}
