% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_sex_adjust.R
\name{age_sex_adjust}
\alias{age_sex_adjust}
\title{Calculate age- and sex-adjusted incidence rates}
\usage{
age_sex_adjust(data, exposure_var, age_var, sex_var, event_var, time_var)
}
\arguments{
\item{data}{A dataframe containting the variables used. Dataframe.}

\item{exposure_var}{Column in data that contains the exposure variable. Numeric or categorical vector.}

\item{age_var}{Column in data that contains the age variable in years. Numerical vector.}

\item{sex_var}{Column in data that contains the sex variable. Binary numerical or categorical vector.}

\item{event_var}{Column in data that contains the event indicator. Binary numerical vector, 0 = no event, 1 = event.}

\item{time_var}{Column in data that contains the follow-up time indicator in days. Numerical vector.}
}
\value{
A dataframe with the following columns
\itemize{
\item \code{Treatment}:  Treatment groups
\item \code{IR}:  Age- and sex-adjusted incidence rate per 100 person year
\item \code{Lower_CI}:  Lower 95\% confidence interval
\item \code{Upper_CI}:  Upper 95\% confidence interval
}
}
\description{
This function uses a poisson model to adjust for age and sex.
It returns a dataframe with estimated incidence rate per 100 person years.
Uses marginaleffects::avg_predictions to obtain marginal population standardized estimates
}
\examples{
colon_death <- colon[colon$etype == 2, ] # Select rows with event indicator and follow-up time for death
age_sex_adjust(colon_death, rx, age, sex, status, time)
age_sex_adjust(colon[colon$etype == 2, ], rx, age, sex, status, time)

}
