% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/applyWeights.R
\name{applyWeights}
\alias{applyWeights}
\title{Apply a set of weights to the columns of an xts object.}
\usage{
applyWeights(w, R)
}
\arguments{
\item{w}{the weight vector. Must have same number of elements as columns in R.}

\item{R}{the xts object of returns or assets.}
}
\value{
an xts object with the weights, w, applied to R.
}
\description{
This function is just a wrapper to the 'sweep' function because the
arguments to sweep are too hard to remember.
}
\examples{
dts <- seq(Sys.Date()-4, Sys.Date(), 1)
returns <- matrix(rnorm(10),ncol = 2) / 100
ret.xts <- xts(returns, dts)
wts <- c(.50, .50)
returns.wtd <- applyWeights(wts, ret.xts)
# check...
returns/returns.wtd
}

