% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextab_functions.R
\name{bioavr_flextab_defaults}
\alias{bioavr_flextab_defaults}
\title{Set default values for flextable objects}
\usage{
bioavr_flextab_defaults()
}
\value{
A function.
}
\description{
This function sets the font family to Calibri, size to 10pts, and border color to black.
}
\examples{
bioavr_flextab_defaults()

}
