% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextab_functions.R
\name{bioavr_group_tab}
\alias{bioavr_group_tab}
\title{Create a grouped flextable object}
\usage{
bioavr_group_tab(flextab, header, footer, default = bioavr_flextab_defaults())
}
\arguments{
\item{flextab}{data. Dataframe.}

\item{header}{header text. String.}

\item{footer}{footer text. String.}

\item{default}{Set flextable defaults, using flextable::set_flextable_defaults(). Function}
}
\value{
A flextable object.
}
\description{
Usually used to create crude and age-, and sex-adjusted incidence rate tables.
Useful for exporting to word using the \code{flextable::save_as_docx()} function.
}
\examples{
colon_death <- colon[colon$etype == 2, ]
dplyr::bind_rows("Crude" = incidrate_crude(colon_death, rx, status, time),
"Age- and sex-adjusted" = age_sex_adjust(colon_death, rx, age, sex, status, time), .id = "id") \%>\%
bioavr_group_tab(header = "Table 2.
Crude and age- and sex-adjusted incidence rates per 100 person-years",
footer = "Estimated using a Poisson model")


}
