% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrapping of correlation coefficients}
\usage{
bootstrap(x, y, n = 10000L, ...)
}
\arguments{
\item{x}{double vector. The observations of the first variable.}

\item{y}{double vector. The observations of the second variable.}

\item{n}{integer. The number of replicates.}

\item{\dots}{further arguments to the function \code{\link[base]{cor}}}
}
\value{
A vector of \code{n} bootstrapping samples.
}
\description{
Function of bootstrapping of correlation coefficients
}
\examples{
## Sample some data
set.seed(123L)
x <- rnorm(50)
y <- x + rnorm(length(x))
cor(x, y)
## bootstrapping
bs <- bootstrap(x, y)
## bootstrap distribution
hist(bs)
## mean and standard-deviation of bootstrap distribution
c(mean(bs), sd(bs))
## two-sided 5\% quantiles
quantile(x = bs, probs = c(0.025, 0.975))

}
