% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cllps.R
\name{cllps}
\alias{cllps}
\title{Collapse elements of a vector}
\usage{
cllps(x, sep = "")
}
\arguments{
\item{x}{A vector to be collapsed.}

\item{sep}{Seperator character used for collapsing.}
}
\value{
A character vector of length one containing the 
collapsed input.
}
\description{
\code{cllps} collapses the elements of a vector
}
\details{
This function is defined as
\code{cllps <- function(x, sep = '') paste(x, collapse = sep)}
}
\examples{
cllps(c(letters), '_')

}
\author{
Dominik Mueller
}
