% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_scale.R
\name{col_scale}
\alias{col_scale}
\title{Center and scale a matrix.}
\usage{
col_scale(x, center = FALSE, scale = FALSE, inplace = FALSE)
}
\arguments{
\item{x}{A numeric matrix, preferable with storage mode double
as the calculations might be faster.}

\item{center}{Logical or a numeric vector (length one or \code{ncol(x)} being
subtracted from the columns of \code{x} for centering.}

\item{scale}{Logical or a numeric vector (length one or \code{ncol(x)} to 
divide the columns of \code{x} for scaling.}

\item{inplace}{Logical. If \code{TRUE}, the matrix is modified in memory,
i.e., no copy is being made. This is faster and saves
memory, but only possible if storage mode is double. If \code{FALSE},
a copy is made, modified and returned.}
}
\value{
Either a modified copy of the original matrix if
\code{inplace = FALSE}, or nothing if \code{inplace = TRUE}.
}
\description{
Center and scale a matrix using given vectors for
centering and scaling.
}
\details{
Arguments to \code{center} and \code{scale} can be either logical
or numeric vectors. If \code{TRUE}, centering is done by column means
and scaling by the sample standard deviation of the columns. If arguments are
numerical, they must be either of length one (recycled to the number of
columns of \code{x}) or of length equal to \code{ncol(x)}. Arguments to
\code{center} and \code{scale} are independent.

This function was specifically desiged for scaling with known \code{center}
and \code{scale} arguments. Note that maximum speed is achieved if the
matrix has storage mode double and is modified in 
memory (\code{inplace = TRUE}).
}
\examples{
x <- matrix(as.double(1:20), nrow = 4)
col_scale(x, center = 4, scale = TRUE)

x <- matrix(rnorm(1e6L), nrow = 1e2L, ncol = 1e4L)
center <- colMeans(x)
scaling <- matrixStats::colSds(x, center = center)
xx <- x + 0.0
xx_ <- x + 0.0

microbenchmark::microbenchmark(times = 10L,
 scale(x, center = center, scale = scaling),
 scale(xx, center = center, scale = scaling),
 col_scale(x, center, scaling),
 col_scale(xx, center, scaling),
 col_scale(xx, center, scaling, inplace = TRUE)
)

}
\author{
Dominik Mueller
}
