% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kipouro_functions.R
\name{csprobdif}
\alias{csprobdif}
\title{Calculate regression standardized differences in cumulative incidence while accounting for the competing risk of death}
\usage{
csprobdif(predprob1, predprob2)
}
\arguments{
\item{predprob1}{One object from prob_list from regstand_cr()}

\item{predprob2}{Another object from prob_list from regstand_cr() to subtract from}
}
\value{
Returns a list with the differences for the various causes. Usually ProbDif1 if what is wanted.
}
\description{
Using methodology described in https://doi.org/10.1002/sim.8209.
Building upon code by Dr. Dimitra-Kleio Kipourou.
}
\examples{
\dontrun{
differences <- csprobdif(prob_list$Male, prob_list$Female)
}
}
