% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut3.R
\name{cut3_quantile}
\alias{cut3_quantile}
\title{Cut3 by quantile}
\usage{
cut3_quantile(tbl, var_name, .labels = NULL, .groups = NULL, .inf = TRUE, ...)
}
\arguments{
\item{tbl}{\link{data.frame}: Database connection or data.frame}

\item{var_name}{\link{character}: variable name}

\item{.labels}{\link{list}: labels for the breaks}

\item{.groups}{\link{character}: name of a groups variable}

\item{.inf}{\link{logical}: indicates if the breaks need to be extended by -Inf and Inf}

\item{...}{argument passed to quantile}
}
\value{
same as \code{tbl} input with \code{var_name} converted to factor by quantiles
}
\description{
Cut3 by quantile
}
\examples{
datos <- data.frame(edad = seq(1:100))
cut3_quantile(datos, "edad")
}
