% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut3.R
\name{cut3_quantiles}
\alias{cut3_quantiles}
\title{Cut3: quantiles}
\usage{
cut3_quantiles(tbl, var_name, weights, probs, ...)
}
\arguments{
\item{tbl}{\link{data.frame}: Database connection or data.frame}

\item{var_name}{\link{character}: variable name}

\item{weights}{\link{character}: name of a numeric variable of weights}

\item{probs}{See \code{Hmisc::\link[Hmisc:wtd.quantile]{wtd.quantile}}}

\item{...}{other arguments passed to quantile functions}
}
\value{
vector of specified quantiles
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\details{
Helper function for convert numeric variables into categorical variables
based on their own quantiles.
}
\examples{
  (tbl <- data.frame(
    edad = 1:10,
    peso = round(runif(10)*10),
    grupo = sample(c("H", "M"), 10, replace = TRUE)))

    cut3(tbl,
         "edad",
         c(0, cut3_quantiles),
         labels = c("1", "2", "3", "4"),
         groups = "grupo",
         weights = "peso",
         probs = c(0.25, 0.5, 0.75, 1))
}
\seealso{
\code{vignette("cut3", package = "Dmisc")}
}
