% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnect.R
\name{dbConnect}
\alias{dbConnect}
\title{Conexión a bases de datos}
\usage{
dbConnect(
  db_user = NULL,
  db_pass = NULL,
  db_sys = "PostgreSQL",
  db_name = "encft",
  db_host = "localhost",
  db_port = 5432,
  k_service_id = "postgre"
)
}
\arguments{
\item{db_user}{\link{character}: nombre de usuario de la base de datos.}

\item{db_pass}{\link{character}: contraseña de la base de datos.}

\item{db_sys}{\link{character}: nombre del sistema de base de datos. Para utilizar el driver correcto.}

\item{db_name}{\link{character}: nombre de la base de datos}

\item{db_host}{\link{character}: dirección del host de la base de datos}

\item{db_port}{\link{numeric}: puerto de la base de datos}

\item{k_service_id}{\link{character}: "keyring service id" si está utilizando keyring
para almacenar sus credenciales.}
}
\value{
conexión a base de datos. Vea \code{DBI::\link[DBI:dbConnect]{dbConnect}}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\details{
Crea una conexión con un sistema de base de datos.
}
\examples{
\dontrun{
conn <- dbConnect()
}
}
\seealso{
\href{https://db.rstudio.com/best-practices/managing-credentials/#encrypt-credentials-with-keyring}{Databases using R}
}
