% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{dend_to_df}
\alias{dend_to_df}
\title{dend_to_df: Transform a dendrogram into a data frame
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dend_to_df(dend)
}
\arguments{
\item{dend}{A dendrogram object.}
}
\value{
A data frame with one row for each leaf in the dendrogram, and columns
representing each level of the dendrogram with values representing the label
for each leaf.
}
\description{
This function takes a dendrogram object and converts it into a data frame.
The resulting data frame has columns representing each level of the dendrogram
and the values representing the label for each leaf.
}
\examples{
\dontrun{
# create example dendrogram
dend <- as.dendrogram(hclust(dist(USArrests), "ave"))

# use the function
df <- dend_to_df(dend)
}
}
