% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{dend_to_list}
\alias{dend_to_list}
\title{Convert a Dendrogram Object to a List
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dend_to_list(dend)
}
\arguments{
\item{dend}{a dendrogram object}
}
\value{
a nested list, where each node in the dendrogram is represented as a
list in the returned list. If a node is a leaf node, it is represented as a
character vector of length one, containing the label of the leaf node.
}
\description{
This function takes as input a dendrogram object, and returns a nested list,
where each node in the dendrogram is represented as a list in the returned list.
If a node is a leaf node, it is represented as a character vector of length one,
containing the label of the leaf node.
}
\examples{
\dontrun{
# Create a dendrogram object
dend <- as.dendrogram(hclust(dist(USArrests), "ave"))

# Convert the dendrogram to a list
dend_to_list(dend)
}

}
