% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Create a table of summaries for an entery dataframe}
\usage{
describe(data, digits = 4, t = TRUE, flextable = FALSE, ...)
}
\arguments{
\item{data}{\link{data.frame}: a dataframe with the data}

\item{digits}{\link{numeric}: number of digits for the numbers}

\item{t}{\link{logical}: indicates if the resulting table is transposed}

\item{flextable}{\link{logical}: indicates if the table is converted to a simple flextable}

\item{...}{other arguments passed to summary function}
}
\value{
a table containing the summary of every variable in the provided dataframe
}
\description{
Create a table of summaries for an entery dataframe
}
\examples{
\dontrun{
describe(cars, flextable = T)
}
}
