% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_ts.R
\name{df_to_ts}
\alias{df_to_ts}
\title{Convert a dataframe to a time series object
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
df_to_ts(data, .freq = NULL, .date = "date", end = NULL)
}
\arguments{
\item{data}{\link{data.frame}: a dataframe with the data}

\item{.freq}{\link{numeric}: the frequency of the time series. If not specified, it is automatically determined by the data}

\item{.date}{\link{character}: the name of the date column in the dataframe}

\item{end}{\link{character}: the last observation of the time series. If not specified, it is automatically determined by the data}
}
\value{
a time series object
}
\description{
Convert a dataframe to a time series object
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
data(fdeaths)
ts_data <- data.frame(time = time(fdeaths), y = fdeaths)
ts <- df_to_ts(ts_data, .freq = 12, .date = "time")
}
}
