% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dftlov.R
\name{dftlov}
\alias{dftlov}
\title{Turn data.frame to list of named vectors.}
\usage{
dftlov(data, values, ids, by)
}
\arguments{
\item{data}{A \code{data.frame}-like object (\code{data.frame}, \code{tibble}, ...).}

\item{values}{A character identifying the column holding the values.}

\item{ids}{A character identifying the column holding the names.}

\item{by}{A character identifying the column holding the factor for splitting.}
}
\value{
A list of name vectors.
}
\description{
Turn a data.frame into a list of named vectors, splitting by another
variable.
}
\examples{
dat <- data.frame(chrom = c(1L, 1L, 2L, 2L, 3L, 3L),
pos = c(1.4, 5.5, 1.2, 6.8, 3.3, 5.4),
id = c('m1', 'm2', 'm3', 'm4', 'm5', 'm6'))
dftlov(dat, values = 'pos', ids = 'id', by = 'chrom')

}
\author{
Dominik Mueller
}

