% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/divRatio.R
\name{divRatio}
\alias{divRatio}
\title{Compute the Diversification Ratio for a portfolio.}
\usage{
divRatio(w = NULL, R, sigma)
}
\arguments{
\item{w}{a vector of weights.}

\item{R}{an xts object or matrix of asset returns.}

\item{sigma}{a covariance matrix for R.}

\item{scale}{the annualization factor for R. Default = 1. This parameter is
    passed through to wtdAvgVol() and portfolioVol().}
}
\value{
the Diversification Ratio (DR) for the portfolio.
}
\description{
This function computes the Diversification Ratio (DR) for a portfolio, given
a weight vector, \code{w}, a set of asset returns, \code{R}, and a
covariance matrix, \code{sigma}. This ratio is a measure of how well
diversified a portoflio is. The square of this number is a proxy for the
number of unique sources of variation exist in a portfolio. The higher the
number, the more diversified the portfolio.
}
\examples{
dts <- seq(Sys.Date()-199, Sys.Date(), 1)
returns <- matrix(rnorm(1000),ncol = 5) / 100
ret.xts <- xts(returns, dts)
cm <- cov(ret.xts)
divRatio(R = ret.xts, sigma = cm)
}
\seealso{
wtdAvgVol, portfolioVol
}

