% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmisc-package.R
\docType{package}
\name{dmisc-package}
\alias{dmisc}
\alias{dmisc-package}
\title{dmisc: Survival Analysis Helper Functions}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MichaelDismorrMD/dmisc}
  \item Report bugs at \url{https://github.com/MichaelDismorrMD/dmisc/issues}
}

}
\author{
\strong{Maintainer}: Michael Dismorr \email{michael.dismorr@ki.se} (\href{https://orcid.org/0000-0002-5652-8774}{ORCID})

}
\keyword{internal}
