% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_arima.R
\name{evaluate_arima}
\alias{evaluate_arima}
\title{Evaluate ARIMA model performance}
\usage{
evaluate_arima(
  model,
  valid = NULL,
  test = NULL,
  xreg = NULL,
  .train_stats = TRUE,
  .resid_tests = TRUE,
  .transf = NULL,
  .metrics = c("rmse", "mae", "mape", "smape"),
  .report = TRUE,
  ...
)
}
\arguments{
\item{model}{An ARIMA model object returned by arima().}

\item{valid}{A time series object with the validation set, if applicable. Default is NULL.}

\item{test}{A time series object with the test set, if applicable. Default is NULL.}

\item{xreg}{An optional regression variable for the model.}

\item{.train_stats}{A boolean to indicate whether to include training metrics (TRUE) or not (FALSE). Default is TRUE.}

\item{.resid_tests}{A boolean to indicate whether to include residual checks (TRUE) or not (FALSE). Default is TRUE.}

\item{.transf}{A function or a list of functions to be applied to the observed and fitted time series. This is useful for handling models that include transformations.}

\item{.metrics}{A vector of metrics to be calculated. Default is c("rmse", "mae", "mape", "smape").}

\item{.report}{A boolean to indicate whether to print the report (TRUE) or not (FALSE). Default is TRUE.}

\item{...}{Not implemented.}
}
\value{
A list with the following elements:
\item{model-summary}{A summary of the model}
\item{training_metrics}{A matrix of training metrics (if .train_stats is TRUE)}
\item{residual_checks}{A list of results of the residual checks (if .resid_tests is TRUE)}
\item{forecast-metrics}{A matrix of forecast metrics (if validation and test sets are supplied)}
\item{pred}{A data frame with the predictions and observed values}
\item{plot_pred}{A ggplot object with a plot of the observed and predicted values}
\item{plot_rmse}{A ggplot object with a plot of the RMSE values}
\item{plot_mae}{A ggplot object with a plot of the MAE values}
\item{plot_mape}{A ggplot object with a plot of the MAPE values}
\item{plot_smape}{A ggplot object with a plot of the sMAPE values}
\item{notes}{A vector of notes about the evaluation process}
}
\description{
Evaluate ARIMA model performance, including training metrics, residual checks, and forecasting performance.
}
\examples{
\dontrun{
model <- arima(lynx, order = c(1, 1, 1))
evaluate_arima(model)
}

}
