% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectationPlot.R
\name{expectationPlot}
\alias{expectationPlot}
\title{Plot a manager track record with ex-ante expection bounds.}
\usage{
expectationPlot(x, exp.sharpe, exp.ann.vol, scale = 12, cols = 1,
  auto.arrange = TRUE, mytheme = theme_bw(base_size = 10))
}
\arguments{
\item{x}{an xts object of returns.}

\item{exp.sharpe}{a vector of ex-ante Sharpe ratios for each of the assets.}

\item{exp.ann.vol}{a vector of annualized volatilities for each asset.}

\item{scale}{the annualization factor for the returns. (eg. 12 = monthly)}

\item{cols}{the number of columns in the graphics window.}

\item{auto.arrange}{set to FALSE if you want to specify the number of columns
in the plot. Otherwise, funds will be auto-arranged in plot window.}

\item{mytheme}{a valide ggplot2 theme. Default is theme_bw.}

\item{...}{additional pass-through paramaters for chart.Timeseries}
}
\value{
a series of expectation plots for each asset in x.
}
\description{
This function takes an xts object of manager returns, an expected ex-ante
Sharpe ratio vector and and ex-ante volatility vector and plots cone charts
for each of the managers. The plots have the expected growth trajectory as
well as one and two standard deviation bands. This helps detect whether
performance of the manager is progressing in line with ex-ante
expectations.
}
\examples{
dts <- seq(Sys.Date()-19, Sys.Date(), 1)
returns <- matrix(rnorm(20),ncol = 1) / 100
ret.xts <- xts(returns, dts)
funds <- xts(matrix(rep(ret.xts,6),ncol = 6),index(ret.xts))
sharpes <- rep(1,6)
vols <- rep(.1,6)
expectationPlot(funds, sharpes, vols)
}

