% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{find_label_node}
\alias{find_label_node}
\title{Find Label Node
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
find_label_node(dend_df, .label, node_num = NULL)
}
\arguments{
\item{dend_df}{A data frame that represents a dendrogram.}

\item{.label}{The label for which to find the node number(s).}

\item{node_num}{The node number to return (optional).}
}
\value{
A numeric value or a vector of numeric values, depending on whether
node_num is provided.
}
\description{
This function takes a dendrogram data frame dend_df, a label .label, and a
node number node_num, and returns the node number corresponding to the provided
label. If node_num is provided, the function returns only the specified node
number. If node_num is not provided, the function returns all node numbers
corresponding to the provided label.
}
\examples{
\dontrun{
dend_df <- dend_to_df(as.dendrogram(hclust(dist(mtcars))))

# Example usage
find_label_node(dend_df, "label1")
find_label_node(dend_df, "label1", node_num = 1)

find_label_node(dend_df, "Mazda RX4")
# returns [1] 5 5 5 5 5

find_label_node(dend_df, "Datsun 710")
# returns [1] 2 2 2 2 2

find_label_node(dend_df, "Mazda RX4 Wag", node_num = 1)
# returns 6
}
}
