% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{find_labels_path}
\alias{find_labels_path}
\title{Find Labels Path
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
find_labels_path(dend_df, .label, .sep = "", .pre = 1)
}
\arguments{
\item{dend_df}{A data frame representing the dendrogram}

\item{.label}{The labels of the nodes to be found}

\item{.sep}{The separator for the path elements}

\item{.pre}{The number of characters to remove from the end of the path elements}
}
\value{
A character vector of the paths to the nodes with the specified labels
}
\description{
Find Labels Path
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
