% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bic.R
\name{get_bic}
\alias{get_bic}
\title{Get Bayes information criterion from a mexhaz model object}
\usage{
get_bic(mod)
}
\arguments{
\item{mod}{a mexhaz model object}
}
\value{
A dataframe with the BIC value
}
\description{
Get Bayes information criterion from a mexhaz model object
}
\examples{
# We will use the pbc dataset from the `survival` package

pbc <- survival::pbc \%>\% dplyr::mutate(time_years = time / 365.241) \%>\%
 dplyr::select(time_years, status, sex, age, albumin)

kts <- quantile(pbc$time_years, probs=c(1/3,2/3)) # Knot positions for the baseline hazard
outcome_mod <- mexhaz::mexhaz(Surv(time_years, status == 1) ~ sex + age + albumin,
                             data = pbc,
                             base = "exp.bs",
                             degree = 3,
                             knots = kts,
                             verbose = 0,
                             print.level = 0)

get_bic(outcome_mod)
}
