% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAngle.R
\name{getAngle}
\alias{getAngle}
\title{Return the angle between to vectors, u and v.}
\usage{
getAngle(u, v)
}
\arguments{
\item{u}{a vector of numeric values. (Must be at least length = 2)}

\item{v}{a vector of numeric values. (Must be at least length = 2)}
}
\value{
the angle between the two vectors u and v.
}
\description{
This function returns the angle between two vectors u and v. If u and v are
orthogonal then the angle will be 90 degrees. This is a helper function for
dMisc::plotCovEllipse(). It can be called directly if needed.
}
\examples{
u <- rnorm(5,0,1)
v <- rnorm(5,0,1.5)
getAngle(u,v)
}

