% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getManagerData.R
\name{getManagerData}
\alias{getManagerData}
\title{Read in a file containing manager or asset data and return a list.}
\usage{
getManagerData(path)
}
\arguments{
\item{path}{a character string specifying the path to the data.}
}
\value{
a list containing:
     \code{mydata} the raw imported data.
     \code{comps.ret} the daily returns of the managers in xts format.
     \code{comps.index} an equally weighted wealth index created from
         \code{comps.ret}.
     \code{comps.index.ret} the returns created from \code{comps.index}.
}
\description{
This function reads in a data file and returns a list. The list contains:
the raw return data, returns of the managers, a wealth index
created by equally weighted average of the managers, returns from that
composite index.
}
\details{
This function is useful if you have a .csv file with
a 'Date' column with dates in "Y-m-d" format and manager or asset return
information in each of the other columns. The heading of each of the non-date
columns should be the asset name. This wrapper just saves you from having
to do all the minor cleaning associated with creating and assembling the
data into a format useful for analysis.
}
\examples{
# Requires dMisc::makeIndex()
library(dMisc)
dts <- seq(Sys.Date()-4, Sys.Date(), 1)
returns <- matrix(rnorm(10),ncol = 2) / 100
ret.df <- data.frame(as.character(dts),returns)
colnames(ret.df) <- c("Date", "asset_1", "asset_2")
tf <- write.csv(ret.df,file = "tf.csv", row.names = F)

mydata <- getManagerData(path = paste(getwd(),"/tf.csv",sep = ""))
}

