\name{goldenRatioColors}
\alias{goldenRatioColors}
\title{Choose n colors using the golden ratio}
\usage{
  goldenRatioColors(n, s = 0.5, v = 1, alpha = 1)
}
\arguments{
  \item{n}{Integer.  The number of colors you want}

  \item{s}{Numeric. Saturation - input into hsv}

  \item{v}{Numeric. Value - input into hsv}

  \item{alpha}{Numeric. The alpha blending value that is
  input into hsv.}
}
\description{
  This choices n colors for you based on breaking up the
  hues according to a sequence generated by the golden
  ratio.
}
\examples{
cols <- goldenRatioColors(5)
plot(1:5, 1:5, col = cols)

n <- 5
xs <- seq(1, n)
plot(c(1, n+1), c(0, 1), type = "n")
rect(xs, 0, xs + 1, 1, col = goldenRatioColors(n))
}

