% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_assign.R
\name{\%...>\%}
\alias{\%...>\%}
\title{Unpack right assignments
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
values \%...>\% names
}
\arguments{
\item{values}{A vector/list/data.frame of values for the assignments.}

\item{names}{A character vector of names for the assignments.}
}
\value{
NULL. Assignments are made in the global environment.
}
\description{
Unpack right assignments
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
c(1, 2) \%...>\% c("num1", "num2")
}
}
