% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_assign.R
\name{\%<...\%}
\alias{\%<...\%}
\title{Unpack assignments
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
names \%<...\% values
}
\arguments{
\item{names}{A character vector of names for the assignments.}

\item{values}{A vector/list/data.frame of values for the assignments.}
}
\value{
NULL. Assignments are made in the global environment.
}
\description{
Unpack assignments
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
# Assignment by position
c("num1", "num2") \%<...\% c(1, 2)

#> num1
#> [1] 1

# Assignment by name
c("num1", "num2") \%<...\% list(num2 = 1, num1 = 2)

#> num1
#> [1] 2
}
}
