% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hitRatio.R
\name{hitRatio}
\alias{hitRatio}
\title{Compute the hit ratio of single or the difference between two track records.}
\usage{
hitRatio(x, y = NULL)
}
\arguments{
\item{x}{a series of observations from a given asset.}

\item{y}{a second track record to compare to x. Default is NULL.}
}
\value{
the percentage of time x is greater than 0 or the percentage of time
    \code{x-y} is greater than zero.
}
\description{
This function allows you to calculate the percentage of observations greater
than zero in the case of a single track record, \code{x}. It also allows you
to pass a second track record. If the seconde track record is passed, then
the function computes the periodic differences and returns the percentage
of observations where the difference \code{x-y} is positive. In the two
track record case, the return value becomes an outperformance frequency
measure.
}
\examples{
dts <- seq(Sys.Date()-49, Sys.Date(), 1)
returns <- matrix(rnorm(100),ncol = 2) / 100
ret.xts <- xts(returns, dts)
hitRatio(ret.xts[,1])
hitRatio(x = ret.xts[,1], y = ret.xts[,2])
}

