\name{holtWinters}
\alias{holtWinters}
\title{Holt winters smoothing no seasonality}
\usage{
  holtWinters(y, alpha = 0.5, beta = 0.5)
}
\arguments{
  \item{y}{Vector. The data to smooth.}

  \item{alpha}{Numeric.  A smoothing parameter}

  \item{beta}{Numeric. A smoothing parameter}
}
\description{
  Holt winters smoothing with no seasonality
}

