% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_dbl.R
\name{\%in_dbl\%}
\alias{\%in_dbl\%}
\title{A version of \code{\link{\%in\%}} working exclusively on numeric data.}
\usage{
x \%in_dbl\% table
}
\arguments{
\item{x}{A numeric vector with the values to be matched.}

\item{table}{A numeric vector with the values to be matched against.}
}
\value{
A vector of the same length as x. A logical vector, indicating if a
match was located for each element of x: thus the values are TRUE or FALSE
and never NA.
}
\description{
\code{\%in_dbl\%} is an intuitive interface analogous to
\code{\link[base]{\%in\%}} as binary operator, which returns a logical vector indicating
if there is a match or not for its left operand
}
\details{
\code{\%in_dbl\%} is defined as

\code{`\%in_dbl\%` <- function(x, table) match_dbl(x, table, nomatch = 0) > 0}
}
\seealso{
\code{\link{match_dbl}}
}
\author{
Dominik Mueller
}
