% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidrate_crude.R
\name{incidrate_crude}
\alias{incidrate_crude}
\title{Calculate crude incidence rates}
\usage{
incidrate_crude(data, exposure_var, event_var, time_var)
}
\arguments{
\item{data}{A dataframe containing the variables used. Dataframe.}

\item{exposure_var}{Column in data that contains the exposure variable. Numeric or categorical vector.}

\item{event_var}{Column in data that contains the event indicator. Binary numerical vector, 0 = no event, 1 = event.}

\item{time_var}{Column in data that contains the follow-up time indicator in days. Numerical vector.}
}
\value{
A dataframe with the following columns
\itemize{
\item \code{Treatment}:  Treatment groups
\item \code{IR}:  Age- and sex-adjusted incidence rate per 100 person year
\item \code{Lower_CI}:  Lower 95\% confidence interval
\item \code{Upper_CI}:  Upper 95\% confidence interval
}
}
\description{
It returns a dataframe with estimated incidence rate per 100 person-years.
Uses marginaleffects::avg_predictions to obtain marginal population standardized estimates
}
\examples{
colon_death <- colon[colon$etype == 2, ]
# The row above selects rows with event indicator and follow-up time for death

age_adjust(colon_death, sex, status, time)

}
