% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexToXTSReturns.R
\name{indexToXTSReturns}
\alias{indexToXTSReturns}
\title{Convert a wealth index to xts returns. This is the inverse of makeIndex().}
\usage{
indexToXTSReturns(x)
}
\arguments{
\item{x}{an xts object containing one or more wealth indices.}
}
\value{
an xts object containing the returns for each wealth index in x.
}
\description{
This function is the inverse of makeIndex. It takes an xts object that
contains a wealth index and converts it to returns.
}
\examples{
dts <- seq(Sys.Date()-4, Sys.Date(), 1)
returns <- matrix(rnorm(5),ncol = 1) / 100
ret.xts <- xts(returns, dts)
x <- makeIndex(ret.xts)
indexToXTSReturns(x)
}

