% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_html.R
\name{insert_html}
\alias{insert_html}
\alias{insertHTML}
\title{Insert HTML file into another
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
insert_html(
  path = deprecated(),
  name = deprecated(),
  file = paste0(path, name),
  source_code = NULL,
  keep_styles = TRUE
)

insertHTML(
  path = deprecated(),
  name = deprecated(),
  file = paste0(path, name),
  source_code = NULL,
  keep_styles = TRUE
)
}
\arguments{
\item{path}{\link{character}: the path where the HTML file exists. Should end with /}

\item{name}{\link{character}: the name of HTML file to insert}

\item{file}{\link{character}: path to HTML file. Should include the file name.}

\item{source_code}{\link{character}: used for highlighting the code into the document.
Also the output is wrapped by a \code{<pre></pre>} tags.}

\item{keep_styles}{\link{logical}: indicates if the styles of the file to be
included are preserved. They only apply to their own content.}
}
\value{
character an HTML character string
}
\description{
Insert HTML file into another
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This function was created to insert HTML file content into
RMarkdown HTML documents. Important considerations:
\itemize{
\item{
The images are converted to base64 strings if needed.
}
\item{
A caption with de file name is added to each image.
}
}
}
\examples{
\dontrun{
insert_html(file = "html/index.html")
}
}
