% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertHTML.R
\name{insertHTML}
\alias{insertHTML}
\title{Insert HTML file into another}
\usage{
insertHTML(path = "./", name, source_code = "matlab")
}
\arguments{
\item{path}{character the path where the HTML file exists. Should end with /}

\item{name}{character the name of HTML file to insert}

\item{source_code}{character used for highlighting the code into the document}
}
\value{
character an HTML character string
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
\lifecycle{experimental}

This function was created to insert MatLab published HTML into
RMarkdown documents. Important considerations:
\itemize{
\item{
Only the body of the document if returned.
}
\item{
The images are converted to base64 strings if needed.
}
\item{
A caption with de file name is added to each image.
}
\item{
The output is wrapped by a \code{<pre></pre>} tag.
}
}
}
\examples{
\dontrun{
insertHTML('html/', 'index.html')
}
}
