% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invwhich.R
\name{invwhich}
\alias{invwhich}
\title{Inverse of \link[base]{which}.}
\usage{
invwhich(indices, length.out, useNames = TRUE)
}
\arguments{
\item{indices}{A vector of indices.}

\item{length.out}{The length of the output}

\item{use.names}{Whether names of indices should be used.}
}
\value{
A logical vector of length \code{length.out} with
\code{TRUE} at the positions specified in \code{indices} and
\code{FALSE} elsewhere.
}
\description{
Converts a vector of indices to a logical vector.
}
\examples{
invwhich(c(5, 3, 7), 10)

}
\author{
flying sheep (\url{http://stackoverflow.com/a/31114463/4477223})
}
