% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_subset.R
\name{is_subset}
\alias{is_subset}
\title{Test if one set is a subset of another.}
\usage{
is_subset(x, y, tolerance = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{A vector containing the first set.}

\item{y}{A vector containing the second set.}
}
\value{
A logical vector of length one. It is \code{TRUE} if all the
values in \code{x} are found in \code{y} and \code{FALSE} otherwise.
}
\description{
\code{is_subset} tests if a set \code{x} is a subset
of another set \code{y}
}
\details{
This functions takes special care of numeric input and
uses \code{\link[dmisc]{\%in_dbl\%}} in this case.
}
\examples{
x <- c(1 - 1e-9, 2 + 1e-10)
y <- c(1, 2, 3)
all(x \%in\% y) # returns FALSE
is_subset(x, y) # returns TRUE

is_subset(letters[1:5], letters)
is_subset(c(1L, 3L), c(1.0, 6.0, 3.0))
}
\author{
Dominik Mueller
}
