% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCOrtho.R
\name{KCOrtho}
\alias{KCOrtho}
\title{Orthogonalize a set of factors with respect to all other independent variables.}
\usage{
KCOrtho(factors)
}
\arguments{
\item{factors}{an xts object containing a time series of independent variables.}
}
\value{
an xts object of factors that have been orthogonalized with respect to
all of the other factors in the set.
}
\description{
This is an implementation of Klien and Chow (2014?) "Orthogonalized Equity
Risk Premia and Systematic Risk Decomposition."
Original code by Greg Murphy, 2016.
}
\details{
This function will return a matrix of factors that have been completed
orthogonalized with respect to every other factor in the set while still
preserving the orginal sizes of the variances. This is helpful in producing
better factor models that have senisble coefficients. Orthogonalizing the
brute force way will yield very tiny variances in the resultant factor set
and leave you with massive coefficients in your model. This routine nicely
solves this problem.
}

