% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingsPlot.R
\name{loadingsPlot}
\alias{loadingsPlot}
\title{Plot coefficients from a linear (or other) model.}
\usage{
loadingsPlot(model, var.names)
}
\arguments{
\item{model}{a model object of the type generated by lm().}

\item{var.names}{a vector of variable names.}
}
\value{
a plot of the coefficients of the model.
}
\description{
This code was written by:
Carlisle Rainey
June 29, 2012
http://www.carlislerainey.com/
}
\details{
...and adapted/modified by Adam Duncan 2015. Any errors are strictly my own.

Plots a nice looking coefficient plot that makes it easy to visualize
factor loadings.
}
\examples{
#Fill this in....

}

