% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maCrossoverSignal.R
\name{maCrossoverSignal}
\alias{maCrossoverSignal}
\title{Compute a moving average cross-over momentum signal.}
\usage{
maCrossoverSignal(x, s = 50, l = 100)
}
\arguments{
\item{x}{an xts time series of prices.}

\item{s}{the number of periods for the shorter moving average (assumed to be
in the same periodicity as x. Meaning, if you pass daily data and s = 20, then
the shorter moving average will be calculated over 20 days. If you pass monthly
data, and s = 20, then the shorter moving average will be calculated over 20
months.}

\item{l}{the number of period for the longer moving average.}
}
\value{
a list containing a time series filled with -1, 0, or +1 signals based on the cross-over patterns of the moving averages, the short moving average series, and
the long moving average series.
}
\description{
This function computes a moving average cross-over signal from a price
series. The series should be in xts format. The function will return
+1 if the shorter moving average of the price series is above the longer
moving average and -1 if below. It will return a 0 if the two moving averages
are the same.
}
\examples{
dts.long <- seq(Sys.Date()-499, Sys.Date(), 1)
ret.long <- matrix(rnorm(500),ncol = 1) * 100
long.xts <- xts(ret.long, dts.long)
ma.sig <- maCrossoverSignal(long.xts)

}

