% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEigenAssets.R
\name{makeEigenAssets}
\alias{makeEigenAssets}
\title{Create a time series of synthetic assets built from principal components.}
\usage{
makeEigenAssets(R, rotation = prcomp(R)$rotation)
}
\arguments{
\item{R}{an xts object containing a time series of asset returns.}

\item{rotation}{an optional matrix containing the eigen vectors derived from the assets.}
}
\value{
an xts object of synthetic 'eigen assets.'
}
\description{
This function takes a set of asset returns and applies the weights from
the eigen vectors of the asset returns. These 'synthetic assets' can then
be compared to factors or manipulated in other useful ways.
}
\details{
The calculation is just the matrix multiplication of \code{R} and the
\code{rotation} matrix of R derived from PCA.
}

