% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNoise.R
\name{makeNoise}
\alias{makeNoise}
\title{Create a set of returns in xts format that follow a specified distribution.}
\usage{
makeNoise(num_assets = 1, num_returns = 60, dates = NULL,
  dist = "rnorm", ...)
}
\arguments{
\item{num_assets}{the number of columns of length \code{n} to create.}

\item{num_returns}{the number of rows you would like returned.}

\item{dates}{the dates you would like to use for the returns. If NULL, the
function will use \code{SysDate()} and go back in time \code{n} days.}

\item{dist}{one of "rnorm", "runif", or "dt".}

\item{...}{other pass-through parameters to the random number generators.}
}
\value{
a set of returns in xts format following the desired distribution.
}
\description{
This function returns a set of returns that follow the distribution passed
to makeNoise by the user. It is a convenience function when you want a set
of random returns that follow a certain distribution. The function is a
wrapper for rnorm etc. that handles all the time series wrangling.
}
\examples{
# Make 5 normally distributed assets...
makeNoise(5)

# Make 5 normally distributed assets, 10 returns each, with sd = .10/sqrt(252), mean = 0)
makeNoise(5, num_returns = 10, mean = 0, sd = .10/sqrt(252))
}

