% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeXTS.R
\name{makeXTS}
\alias{makeXTS}
\title{Convert a dataframe of returns to xts format.}
\usage{
makeXTS(x, format = "\%Y-\%m-\%d", names = colnames(x)[2:ncol(x)])
}
\arguments{
\item{x}{a data frame of returns with the first column a date column labeled
'Date'.}

\item{format}{a recognized date format. Defaults to "Y-m-d" format.}

\item{names}{a vector of names for each of the columns, excluding the date
column. Defaults to the colnames of the dataframe x[2:ncol(x)].}
}
\value{
an xts object based on x.
}
\description{
Supposing you have read in a data file that has the first column containing
dates followed by subsequent columns of asset returns, this function will
convert that dataframe into an xts object suitable for analysis with xts based
analysis routines, like PerformanceAnalytics.
}
\examples{
mydata <- read.csv("mydata.csv",sep = ",")
mydata.xts <- makeXTS(mydata)
}

