% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_list.R
\name{nested_list}
\alias{nested_list}
\title{Crease nested lists.}
\usage{
nested_list(len, fill = NA)
}
\arguments{
\item{len}{A integer vector indicating the number of lists a different
hierarchical levels of the nested list. For instance, \code{len = c(2, 3)}
produces a nested list with 2 lists at the first level and 3 at the 
second level, i.e., 2 * 3 entries in total.}

\item{fill}{Value to fill in at the lowest hierarchical level.}
}
\value{
A nested list initialized with \code{fill}.
}
\description{
Create deeply nested list and initialize the values at
the lowest level.
}
\examples{
nested_list(len = c(2, 4, 3), fill = NA)

}
\author{
Dominik Mueller (\email{dominikmueller64@yahoo.de})
}

