% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_to_date.R
\name{numeric_month}
\alias{numeric_month}
\title{Convert a month variable to a numeric format}
\usage{
numeric_month(tbl, month)
}
\arguments{
\item{tbl}{a data.frame with the month variable}

\item{month}{the name or position of the month variable}
}
\value{
tbl a new data.frame with the month variable in numeric format
}
\description{
Convert a month variable to a numeric format
}
\examples{
\dontrun{
  numeric_month(tbl, "month")
}
}
