% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_to_date.R
\name{numeric_quarter}
\alias{numeric_quarter}
\title{Convert a quarter variable to a numeric format}
\usage{
numeric_quarter(tbl, quarter)
}
\arguments{
\item{tbl}{a data.frame with the quarter variable}

\item{quarter}{the name or position of the quarter variable}
}
\value{
tbl a new data.frame with the quarter variable in numeric format
}
\description{
Convert a quarter variable to a numeric format
}
\examples{
\dontrun{
 numeric_quarter(tbl, "quarter")
}
}
