% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot_encoding.R
\name{one_hot_encoding}
\alias{one_hot_encoding}
\title{Apply One-Hot-Encoding to factor variables
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
one_hot_encoding(tbl, variables = NULL, drop_one = FALSE, ...)
}
\arguments{
\item{tbl}{\link{data.frame}: dataset with the variables}

\item{variables}{variables to apply one-hot-encoding. If \link{NULL} all factor
variables are processed.}

\item{drop_one}{\link{logical}: especify if one of categories is drop in final dataset}

\item{...}{other arguments passed to \link[tidyr]{pivot_wider}}
}
\value{
a dataset with new binary variables
}
\description{
Apply One-Hot-Encoding to factor variables
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
  warpbreaks
  one_hot_encoding(warpbreaks)
}
}
