% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalize.R
\name{orthogonalize}
\alias{orthogonalize}
\title{Orthogonalize a set of factors with respect to some independent variable.}
\usage{
orthogonalize(mkt_var, f, reconstitute = T, demean = F, inc_alphas = T)
}
\arguments{
\item{mkt_var}{and xts vector representing the independent variable.}

\item{f}{an xts object of factor variables.}

\item{reconstitute}{a flag to indicate wheter the mkt_var should be
returned along with the orthoganalized factors. Default is TRUE.}
}
\value{
an xts object of factors that have been orthogonalized with respect
    to \code{mkt_var}
}
\description{
This function will regress the factor set f against 'mkt_var' in a series
of single factor regressions. The function will return the an xts object
that contains the intercept + residuals from regressing each of the columns
of \code{f} on the \code{mkt_var}. If \code{reconstitute} is TRUE, then
\code{mkt_var} will be returned in column 1. If \code{reconstitue = FALSE},
then \code{mkt_var} is not returned with the others.
}

