% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalize_all.R
\name{orthogonalize_all}
\alias{orthogonalize_all}
\title{Orthogonalize a set of factors with respect to all other independent variables.}
\usage{
orthogonalize_all(F)
}
\arguments{
\item{f}{an xts vector representing the factor set.}
}
\value{
an xts object of factors that have been orthogonalized with respect
    to all the other factors in the set.
}
\description{
This function will regress each elemen of the factor set against every other
element in the factor set. This produces a factor set where each factor in the set
contains no information about the other factors in the set.
}

