% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaFlip.R
\name{pcaFlip}
\alias{pcaFlip}
\title{Flip the signs on eigen vectors so that PC1 is positive.}
\usage{
pcaFlip(x)
}
\arguments{
\item{x}{a matrix of eigen vectors.}
}
\value{
a matrix containing flipped eigen vectors.
}
\description{
This function ensures that PC1 is always positive. Sometimes, the rotation
matrix form PCA returns a set of eigen vectors where PC1 is mostly negative
signs. That's usually undesirable for most financial applications. This
function just flips things around so they are more sensible from a finance
perspective.
}

