% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{pin_get_or_create}
\alias{pin_get_or_create}
\title{Retrieve or Create a Pin}
\usage{
pin_get_or_create(.data, .board, .name, type = "csv", ...)
}
\arguments{
\item{.data}{The data to be pinned if a new pin is to be created.}

\item{.board}{The board where the pin is located or should be created.}

\item{.name}{The name of the pin to be retrieved or created.}

\item{type}{The type of data being pinned, defaults to 'csv'.}

\item{...}{Additional arguments to be passed to \code{pins::pin_write}.}
}
\value{
The version of the pin that was retrieved or created.
}
\description{
This function attempts to retrieve a specified pin from a board.
If the pin doesn't exist or the data is different, it creates a new pin.
}
\examples{
\dontrun{
  board <- pins::board_register_local()
  data <- mtcars
  pin_version <- pin_get_or_create(data, board, "mtcars_pin")
}
}
