% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCovEllipse.R
\name{plotCovEllipse}
\alias{plotCovEllipse}
\title{Plot covariance elispe between a and b with marginal densitites in margin.}
\usage{
plotCovEllipse(a, b)
}
\arguments{
\item{a}{a vector of numeric values}

\item{b}{a vector of numeric values with the same length as a.}
}
\value{
a scatterplot with the covariance ellipse of a and b displayed.
}
\description{
This function makes a scatterplot of a and b.  It plots the sample
histograms in the margins. It also overlays the kernal densities for a and b.
It then overlays the covariance ellipse for a and b. This function is useful
for detecting whether or not a and b are invariants.
}
\examples{
library(dMisc) # requires access to function getAngle()
a <- rnorm(500, mean = 0 , sd = 1)
b <- rnorm(500, mean = 0, sd  = 1.5)
plotCovEllipse(a,b)

}

