% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioVol.R
\name{portfolioVol}
\alias{portfolioVol}
\title{Compute portfolio volatility from a set of returns.}
\usage{
portfolioVol(w = NULL, R, sigma, scale = 1)
}
\arguments{
\item{w}{a vector of weights. If not supplied, function will assign equal
weights. Default value is \code{NULL}. (ie. equal weights)}

\item{R}{an xts or matrix of asset returns.}

\item{sigma}{a covariance matrix for the asset returns.}

\item{scale}{an annualization factor for the returns if desired. Default = 1.}
}
\value{
a numeric value representing the portfolio volatility.
}
\description{
This function just calculates the portfolio volatility given a set of asset
returns, \code{R}, a weight vector, \code{w}, and a covariance matrix,
\code{sigma}. A \code{scale} value can be supplied which is the annualization
factor for the returns if annualized volatility is desired. The default is
1 (ie. no annualization). Supply 12 for monthly data, 252 for daily, and
so on.
}
\details{
If no weight vector is supplied, the function will calculate equal weights
for the portfolio assets.
}
\examples{
dts <- seq(Sys.Date()-199, Sys.Date(), 1)
returns <- matrix(rnorm(1000),ncol = 5) / 100
ret.xts <- xts(returns, dts)
cm <- cov(ret.xts)
portfolioVol(R = ret.xts, sigma = cm)

}
\seealso{
wtdAvgVol, divRatio
}

