% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predictionError.R
\name{predictionError}
\alias{predictionError}
\title{Calculate the RMSE (Root Mean Squared Error) of a prediction model.}
\usage{
predictionError(y, yhat)
}
\arguments{
\item{y}{the response or actual data series.}

\item{yhat}{the fitted series.}
}
\value{
the RMSE or prediction error  = sqrt(mean((y-yhat)^2))
}
\description{
Given an actual series, \code{y} and a fitted series, \code{yhat}, this
function will return the RMSE, what I'm calling 'Prediction Error'. This is
useful when comparing out-of-sample results from the modeling process across
different fitting methods.
}
\examples{
x <- rnorm(15)
y <- x + rnorm(15)
fit <- lm(y ~ x)
new <- data.frame(x = rnorm(15))
yhat <- predict(fit, newdata = new)
predictionError(y,yhat)
}

