% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_bar.R
\name{progress_bar}
\alias{progress_bar}
\title{Html Progress Bar
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}`}
\usage{
progress_bar(progress, ...)
}
\arguments{
\item{progress}{\link{numeric} The progress value in percentage (0-100).}

\item{width}{\link{numeric} The width of the progress bar.}

\item{height}{\link{numeric} The height of the progress bar.}

\item{suffix}{\link{character} The text to display after the progress value (e.g. "\%").}

\item{font_family}{\link{character} The font family to use for the text.}

\item{font_size}{\link{numeric} The font size of the text, in pixels.}

\item{styles}{\link{character} A string with inline styles to apply to the SVG element.}
}
\value{
an html object that displays a progress bar
}
\description{
Create a html progress bar that displays the progress value in 4 different states, depending on its percentage:
\itemize{
\item{0-25\%}
\item{25-50\%}
\item{50-75\%}
\item{75-100\%}
}
}
\examples{
progress_bar(77)
}
