% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGCFID.R
\name{readGCFID}
\alias{readGCFID}
\title{readGCFID}
\usage{
readGCFID(path = "./", animals)
}
\arguments{
\item{path}{the path that conatins the recordings (has to end in a trailing
/)}

\item{animals}{a character vector containing the names of the recordings /
animals}
}
\value{
a data.frame
}
\description{
Read data recorded with a Thermo GCFID. Data has to be manually
exported as .txt file (export chromatogram to the clipboard and pasted into a
.txt file)
}
\details{
\strong{animals} the elements need to exist as folder in 'path', the
  individual recordings need to be named 'animal\strong{_}odor\strong{-}concentration.txt'
}

