% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kipouro_functions.R
\name{regstand_cr}
\alias{regstand_cr}
\title{Obtain regression standardized cumulative incidence while
accounting for the competing risk of death}
\usage{
regstand_cr(
  data,
  exposure,
  outcome_mod = NA,
  death_mod = NA,
  maxt = NA,
  fragm = NA
)
}
\arguments{
\item{data}{The dataframe used to generate the models.}

\item{exposure}{The column with the variable of interest. Should be a factor.}

\item{outcome_mod}{A \code{mexhaz} mod with the outcome of interest.}

\item{death_mod}{A \code{mexhaz} model with death as outcome.}

\item{maxt}{Max follow-up time in year to do predictions on.}

\item{fragm}{How many splits that should be made from 0 to maxt}
}
\value{
A list containing the following
\itemize{
\item A dataframe with the estimates that can be used for plotting
\item Lists for each level of \code{exposure} resulting from a call to \code{predict_prob()}
}
}
\description{
Obtain regression standardized cumulative incidence while
accounting for the competing risk of death
}
\examples{
# We will use the pbc dataset from the `survival` package

pbc <- survival::pbc \%>\% dplyr::mutate(time_years = time / 365.241)

kts <- quantile(pbc$time_years, probs=c(1/3,2/3)) # Knot positions for the baseline hazard

maxt <- floor(max(pbc$time_years)) # Analyze up to maximum years

fragm <- maxt * 4 # Analyze at each quarter of follow-up

outcome_mod <- mexhaz(Surv(time_years, status == 1) ~ sex + age + albumin + spiders,
data = pbc,
base = "exp.bs",
degree = 3,
knots = kts,
verbose = 0,
print.level = 0)

death_mod <- mexhaz(Surv(time_years, status == 2) ~ sex + age + albumin + spiders,
data = pbc,
base = "exp.bs",
degree = 3,
knots = kts,
verbose = 0,
print.level = 0)

regstand_result <- regstand_cr(pbc, sex, outcome_mod, death_mod, maxt, fragm)

}
