% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapper.R
\name{remapper}
\alias{remapper}
\title{Create a function for re-mapping values.}
\usage{
remapper(old, new, na_when_missing = TRUE)
}
\arguments{
\item{old}{Vector. The old values.}

\item{new}{Vector. The new values.}

\item{na_when_missing}{Logical. If a value cannot be found, should \code{NA} be used or
value queried for? (defaults to \code{TRUE}).}
}
\value{
A vector with re-mapped values.
}
\description{
This is a function factory for creating a function that maps old values
to new ones.
}
\details{
The factory internally uses a \code{\link[hashmap]{hashmap}} for mapping old to new
values. For the possible data types of \code{old} and \code{new}, consult the
documentation of \code{\link[hashmap]{hashmap}} (Here, \code{keys} corresponds to \code{old}
and \code{values} to \code{new}).
}
\examples{
old <- letters
new <- toupper(old)
rmp <- remapper(old = old, new = new)
x <- c('a', 'qrz', 'd')
rmp(x)
remapper(old = old, new = new, na_when_missing = FALSE)(x)

}
\author{
Domink Mueller (\email{dominikmueller64@yahoo.de})
}
