% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repCol.R
\name{repCol}
\alias{repCol}
\title{Form a matrix based on repeating a column x, n times.}
\usage{
repCol(x, n)
}
\arguments{
\item{x}{the column you want replicated in matrix form.}

\item{n}{the number of columns you would like in your matrix.}
}
\value{
a matrix with column x repeated n times.
}
\description{
This function replicates the column x, n times and returns a matrix.
}
\examples{
a <- c(1,2,3,4,5)
repCol(a, 10)

}
\seealso{
repRow
}

