% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repRow.R
\name{repRow}
\alias{repRow}
\title{Repeat the rows of a matrix n times.}
\usage{
repRow(x, n)
}
\arguments{
\item{x}{the row you want replicated in matrix form.}

\item{n}{the number of rows you would like in your matrix.}
}
\value{
a matrix with x repeated n times.
}
\description{
This function replicates the rows of matrix n times.
}
\examples{
a <- c(1,2,3,4,5)
repRow(a, 10)

}
\seealso{
repCol
}

