% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curl.R
\name{req_to_curl}
\alias{req_to_curl}
\title{Convert a httr2 request to a curl command}
\usage{
req_to_curl(req)
}
\arguments{
\item{req}{A httr2 request object.}
}
\value{
A character string containing the curl command.
}
\description{
This function takes a httr2 request object and converts it to a curl command
that can be run in the terminal.
}
\examples{
\dontrun{
req <- request("https://httpbin.org/post") \%>\%
  req_body_json(list(a = 1, b = 2), encode = "json") \%>\%
  req_headers("Content-Type" = "application/json")
 
req_to_curl(req) |>
  cat()
#> 'curl -X POST \
#>   -H "Content-Type: application/json" \
#>   -d '{"a":1,"b":2}' \
#>   "https://httpbin.org/post"'
}
 
 @import httr2
 
 @export
}
