% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvel.R
\name{rmvel}
\alias{rmvel}
\title{Remove elements from a vector or list}
\usage{
rmvel(x, ...)
}
\arguments{
\item{x}{the original data}

\item{...}{elements that should be removed}
}
\value{
A vector or list with the specified elements being removed.
}
\description{
\code{rmvel} removes specified elements from a
vector or list.
}
\details{
This function is primariy intended for integers and
character literals. Doubles are not specifically treated.
}
\examples{
rmvel(list(5, 7, 'hi'), 'hi', 7, 90)

}
\author{
Dominik Mueller
}

