% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrcombn.R
\name{rrcombn}
\alias{rrcombn}
\title{All round robin combinations.}
\usage{
rrcombn(x, closed = TRUE, simplify = TRUE)
}
\arguments{
\item{x}{a vector or list}

\item{closed}{Whether the round robin design is closed, i.e., the last element is
paired up with the first.}

\item{simplify}{Whether the result should be simplified to a matrix.}
}
\value{
A list or matrix containing all combinations.
}
\description{
Returns all pairs of elements occuring in a round robin design.
}
\examples{
rrcombn(letters) 

}
\author{
Dominik Mueller
}
