% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsign.R
\name{rsign}
\alias{rsign}
\title{Sample random signe}
\usage{
rsign(n, p = 0.5)
}
\arguments{
\item{n}{number of samples to draw}

\item{p}{probability that a positive sign is drawn.}
}
\value{
A numeric vector with elements \code{1} and \code{-1}.
}
\description{
Sample a random sign according to a specified probability.
}
\examples{
rsign(10L)
rsign(10L, p = 0.95)

}
\author{
Dominik Mueller
}
