% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubNames.R
\name{scrubNames}
\alias{scrubNames}
\title{Scrub the fund or asset names from an xts object and replace with 'Fund.x'.}
\usage{
scrubNames(x)
}
\arguments{
\item{x}{an xts object to be scrubbed.}
}
\value{
an xts object with the names replaced by 'Fund.x' where x is a
    number ranging from 1 to ncol(x).
}
\description{
Sometimes it is necessary to cover up the names of the funds when
distributing analyses to colleagues. This little function takes care of that
problem.
}
\examples{
dts <- seq(Sys.Date()-199, Sys.Date(), 1)
returns <- matrix(rnorm(1000),ncol = 5) / 100
ret.xts <- xts(returns, dts)
colnames(ret.xts) <- rep(c("Real Name"),ncol(ret.xts))
ret.nonames <- scrubNames(ret.xts)
head(ret.nonames)

}

