% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdAutoCor.R
\name{sdAutoCor}
\alias{sdAutoCor}
\title{Calculate an approximate standard deviation of returns that accounts for
auto-correlation in the returns.}
\usage{
sdAutoCor(x, L = 12)
}
\arguments{
\item{x}{an xts object containing a time series of asset returns.}

\item{L}{the maximum number of lags to consider. Default is 12.}
}
\value{
a list of four calculations described above.
}
\description{
This function is an implementation of the calculations found in
Kinlaw, Kritzman, and Turkington have who published two companion papers
called “The Divergence of High- and Low-Frequency Esitmation: Causes and
Consequences” and "The Divergence of High- and Low-Frequency Esitmation:
Implications for Performance Measurement."
}
\details{
The function returns a set of calculations:
 1. The no-auto-correlation adjusted standard deviation \code{sigma*sqrt(t)}
 2. The auto-correlation adjusted measure : \code{sigma*(sqrt(t) + cors)}
 3. The difference between the two measures.
 4. The ratio of the two measures.
}

