% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_study_folder.R
\name{select_study_folder}
\alias{select_study_folder}
\title{Select study root directory to chose where to load and save files from}
\usage{
select_study_folder()
}
\value{
A string with the path to the study folder
}
\description{
Select study root directory to chose where to load and save files from
}
\examples{
my_root_path <- select_study_folder()
# Chose one number corresponding the the folder you want to use and press enter
# That is now stored in the global variable my_root_path

}
