% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpSignal.R
\name{simpSignal}
\alias{simpSignal}
\title{Compute a buy/sell momentum signal based on the specifed look back period.}
\usage{
simpSignal(x, lookback = 12)
}
\arguments{
\item{x}{an xts object of prices.}

\item{lookback}{the length of the look back period over which to evalute
the momentum signal.}
}
\value{
an xts object of discrete momentum signals drawn from the set c(-1,1),
    where -1 indicates "short" or "sell" and 1 indicates "long" or "buy".
}
\description{
This function will return a vector of buy/sell signals based on whether
or not the current value of \code{x} is above or below the price of
\code{x}. If the current price is above, the return entry is 1 otherwise
the return entry will be zero. The return vector will be of length
\code{x} - \code{lookback}. This function aids in building a smooth
momentum signal by calling simpSignal() with many values of \code{lookback}
and averaging the results.
}
\examples{
library(dMisc)
dts <- seq(Sys.Date()-99, Sys.Date(), 1)
returns <- matrix(rnorm(100),ncol = 1) / 100
ret.xts <- xts(returns, dts)
s <- simpSignal(ret.xts, lookback = 5)
plot(s, type = "l")

}

