% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_sie.R
\name{solve_sie}
\alias{solve_sie}
\title{Solve selection index equations for animal model}
\usage{
solve_sie(y, A, lambda, ps_only = TRUE, check = TRUE)
}
\arguments{
\item{y}{double vector. Phenotypic values (only for training individuals).}

\item{A}{double matrix. A semi-positive-definite matrix (pedigree- or marker-based).}

\item{lambda}{double. The ratio of environmental to genetic variance component.}

\item{ps_only}{boolean. Should predictions only be give for the
prediction set?}

\item{check}{boolean. Should checks on the input be performed?}
}
\value{
A list with two components. The first components hold the BLUE of the general mean,
the second component the BLUPs of the genomic or breeding values.
}
\description{
This is a function for getting BLUPs of genomic/breeding values from the
             animal model using the selection index approach.
}
\details{
This function has two requirements on the sorting of \code{y} and \code{A}:
(i) any individuals to be predicted (having no phenotype) comes before any
training individual with p henotype and (ii) the order of training
individuals in \code{y} is the same as in \code{A}.
}
\author{
Dominik Mueller
}
