% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorted_sample.R
\name{sorted_sample}
\alias{sorted_sample}
\title{Draw sorted sample.}
\usage{
sorted_sample(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{Either a vector of one or more elements from which to choose,
    or a positive integer.  See \sQuote{Details.}}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{Should sampling be with replacement?}

\item{prob}{A vector of probability weights for obtaining the elements
    of the vector being sampled.}
}
\value{
A vector with the sample according to the original order.
}
\description{
Draw a sorted random sample from a vector, conserving
the original order of the elements.
}
\details{
Note that the sample is not actually sorted, but reorder by
matching it back with \code{\link[base]{match}} to the original vector.
}
\examples{
sorted_sample(letters, 5L)

sorted_sample(letters[seq_len(3L)], 10L, replace = TRUE)

}
\author{
Dominik Mueller
}

