% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorted_sample_int.R
\name{sorted_sample_int}
\alias{sorted_sample_int}
\title{Draw sorted sample of integers.}
\usage{
sorted_sample_int(n, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{n}{a positive number, the number of items to choose from.  See
    \sQuote{Details.}}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights for obtaining the elements
    of the vector being sampled.}
}
\value{
A vector of length \code{size} with an increasingly sorted sample
 of integers between \code{1} and \code{n}.
}
\description{
Draw a sorted sample of integers.
}
\examples{
sorted_sample_int(10L, 5L)

}
\author{
Dominik Mueller
}
