% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_str.R
\name{split_str}
\alias{split_str}
\title{Split elements of a string.}
\usage{
split_str(x, split)
}
\arguments{
\item{x}{
    character vector, each element of which is to be split.  Other
    inputs, including a factor, will give an error.
  }

\item{split}{String to split \code{x}.}
}
\value{
A list of the same length as \code{x}, the i-th element of which contains the vector 
of splits of \code{x[i]}.
}
\description{
Split the elements of a string \code{x} into substrings accroding
to (fixed, no regex) matches of \code{split}.
}
\details{
This function is equivalent to \code{strsplit(x, split, fixed = TRUE)[[1L]]}.
}
\examples{
x <- '2_3_4'
split_str(x, '_')
}
\author{
Dominik Mueller
}
\seealso{
\code{\link[base]{strsplit}}
}

