% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_view.R
\name{st_view}
\alias{st_view}
\title{st_view
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
st_view(sf)
}
\arguments{
\item{sf}{An object of class sf (Simple Features).}
}
\value{
Returns an invisble sf object.
}
\description{
The function provides a simple way to view the data of a sf object without
the geospatial component.
}
\examples{
\dontrun{

# Create an sf object
sf_obj <- sf::st_as_sf(mtcars, coords = c("wt", "mpg"), crs = 4326)

st_view(sf_obj)
}
}
