% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl_plot.R
\name{stl_plot}
\alias{stl_plot}
\title{STL Plot}
\usage{
stl_plot(df, var_name, date_var = "date", .freq = 4)
}
\arguments{
\item{df}{A data frame with time series data.}

\item{var_name}{The name of the variable in df to be decomposed.}

\item{date_var}{The name of the date/time variable in df.}

\item{.freq}{The frequency of the time series.}
}
\value{
A plot of the decompositions of the time series.
}
\description{
This function performs STL (Seasonal-Trend Decomposition of Time Series)
decomposition on a time series data frame and plots the resulting trend,
seasonal, and residual components.
}
\examples{

data(economics, package = "ggplot2")
df <- economics

# Plot decomposed time series
stl_plot(df, var_name = "unemploy", date_var = "date", .freq = 12)
}
