% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_c2.R
\name{str_c2}
\alias{str_c2}
\title{str_c2: Concatenate Strings with Custom Last Delimiter}
\usage{
str_c2(..., sep = "", collapse = NULL, last_coll = " y ")
}
\arguments{
\item{...}{strings to be concatenated.}

\item{sep}{delimiter to be used to concatenate the strings.}

\item{collapse}{default delimiter used in base str_c function.}

\item{last_coll}{delimiter to be used for the last pair of strings.}
}
\value{
a single string that is the concatenation of all the input strings.
}
\description{
This function concatenates multiple strings into a single string using a specified delimiter and provides the option to change the delimiter for the last pair of strings.
}
\examples{
str_c2("A", "B", "C", "D", sep = ", ")
str_c2("A", "B", "C", "D", sep = ", ", last_coll = " and ")
}
