% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rows_ext.R
\name{summary_rows_ext}
\alias{summary_rows_ext}
\title{gt summary rows with external values}
\usage{
summary_rows_ext(.x, values, start_pos = 1)
}
\arguments{
\item{.x}{The column values gt passes to the function.}

\item{values}{A vector of values to use in the summary rows.
The values are used in the order they are provided.}

\item{start_pos}{The position in the \code{values} vector to start using values.}
}
\value{
A value from the \code{values} vector for each column.
}
\description{
gt summary rows with external values
}
\examples{
\dontrun{
ext_total <- c(100, 200)
gt(mtcars) \%>\%
  summary_rows(
    groups = vars(cyl),
    columns = vars(mpg, disp),
    fns = list(
      "Total" = ~summary_rows_ext(.x, ext_total)
    )
  )
}
}
