% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{theme_dahaniel}
\alias{theme_dahaniel}
\title{Dahaniel ggplot2 theme}
\usage{
theme_dahaniel(base_size = 7, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{font family used}
}
\value{
a ggplot2 theme based on solarized 2
}
\description{
Dahaniel ggplot2 theme
}

