% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tickersToXTS.R
\name{tickersToXTS}
\alias{tickersToXTS}
\title{Convert a character list of tickers to an xts object.}
\usage{
tickersToXTS(tickers, na.omit = TRUE)
}
\arguments{
\item{tickers}{a character vector of symbols.}

\item{na.omit}{a flag indicating whether NAs should be removed.}
}
\value{
a single xts object containing all of the prices in \code{tickers}.
}
\description{
Takes a character list of tickers and returns an xts object of the
variables. Useful for creating objects after a call to dMisc::getData().
For example, typically after calling dMisc::getData() (a wrapper for
quantmod::getSymbols()) you will end up with OHLC objects in .GlobalEnv.
This function grabs all the variables and uses zoo::merge.zoo to bind
them together in a single object. Return type is set to xts via as.xts().
}
\examples{
tickers <- c("IBM","JPM")
quantmod::getSymbols(tickers)
data <- Cl(tickersToXTS(tickers)) # just closing prices
}

