% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topN.R
\name{topN}
\alias{topN}
\title{Average value of the top elements in each row of a matrix.}
\usage{
topN(x, nth)
}
\arguments{
\item{x}{a numeric matrix}

\item{nth}{an integer giving the desired number of top elemens}
}
\value{
A numeric vector containing the respective means of the top \code{nth}
elements in each row.
}
\description{
Takes a matrix and computes the arithmetic mean of the top elemens for
each single row.
}
\examples{
(x <- apply(matrix(seq_len(16), ncol = 4), 2L, sample))
topN(x, 2L)

# benchmarking with a pure R solution
x <- matrix(rnorm(1000L ^ 2L), ncol = 1000L)
microbenchmark::microbenchmark(times = 1L,
  a <- apply(x, 1L, function(x) mean(head(sort(x, decreasing = TRUE), 20L))),
  b <- topN(x, 20L)
)
all(dplyr::near(a, b))

}
\author{
Dominik Mueller
}
