% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_split}
\alias{train_split}
\title{Get train set from a dataset split
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
train_split(tts)
}
\arguments{
\item{tts}{\link{list}: dataset split}
}
\value{
\link{data.frame}: train set
}
\description{
Get train set from a dataset split
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
warpbreaks
tts <- Dmisc::train_test_split(tbl, test_size = 0.2)
train <- train_split(tts)
}
}
