% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{Split data into train and test sets
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
train_test_split(data, prop = 3/4, val_split = FALSE, ...)
}
\arguments{
\item{data}{\link{ts or data.frame}: the dataset to be split}

\item{prop}{\link{numeric}: the proportion of the data to be assigned to the training set. If prop is less than 1, it is treated as a fraction of the total number of observations in the data. If prop is greater than or equal to 1, it is treated as the exact number of observations to be assigned to the training set.}

\item{val_split}{\link{logical}: whether or not to create a validation set. If val_split is TRUE, a validation set is created from the test set.}

\item{...}{\link{named arguments}: additional arguments passed to rsample::initial_split for non-time series data.}
}
\value{
\link{list}: a list with train, test, and valid (if val_split is TRUE) data sets.
}
\description{
This function splits a dataset into train and test sets. The function can handle both time series and regular data frames. For time series data, the function splits the data based on time, while for regular data frames, it uses the rsample package to split the data randomly.
}
\examples{
\dontrun{
# Time series data
library(lubridate)
library(tidyverse)
ts_data <- ts(data = sin(seq(from = 0, to = 2 * pi, length.out = 100)), start = ymd("2022-01-01"), frequency = 12)
train_test_split(ts_data, prop = 0.8)

# Non-time series data
data(mtcars)
train_test_split(mtcars, prop = 0.8)
}
}
