% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_concat.R
\name{ts_concat}
\alias{ts_concat}
\title{Concatenate Time Series Objects}
\usage{
ts_concat(x)
}
\arguments{
\item{x}{A list of time series objects or a time series object.}
}
\value{
A time series object.
}
\description{
The function ts_concat concatenates time series objects into one time series
object. If the input is a list of time series objects, then it first converts
each time series object to a data frame using the as.data.frame function, and
then concatenates the resulting data frames using the dplyr::bind_rows function.
If the input is not a list, it converts it to a time series object using the
ts function.
}
\examples{
data(airlines)

# Concatenating a list of time series objects
x <- list(airlines[1:5], airlines[6:10])
ts_concat(x)
}
