% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dummy.R
\name{ts_dummy}
\alias{ts_dummy}
\title{Time Series Dummy Variable
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ts_dummy(dates, start = NULL, end = NULL, frequency = NULL, ts = NULL, ...)
}
\arguments{
\item{dates}{dates to set ones}

\item{start}{time series start date}

\item{end}{time series end date}

\item{frequency}{time series frequency}

\item{...}{other arguments passed to ts function}
}
\value{
a time series dummy variable
}
\description{
Time Series Dummy Variable
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{

tsd1 <- ts_dummy(
  dates = 2005,
  start = 2000,
  end = 2021
)
tsd1

tsd2 <- ts_dummy(
  dates = c(2005, 3),
  start = 2000,
  end = c(2021, 4),
  frequency = 4
)
tsd2

tsd3 <- ts_dummy(
  dates = data.frame(c(2005, 2006), c(8, 11)),
  start = 2000,
  end = c(2021, 12),
  frequency = 12
)
tsd3

}
