% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_ggplot.R
\name{ts_ggplot}
\alias{ts_ggplot}
\title{ts_ggplot: Time Series Visualization with ggplot2
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ts_ggplot(ts, .facet = TRUE, .hline = NULL, .ncol = 1, .select = NULL)
}
\arguments{
\item{ts}{A time series object}

\item{.facet}{a logical value. When TRUE (default), the plot will have a facet for each series.}

\item{.hline}{a numeric value to add a horizontal line to the plot.}

\item{.ncol}{the number of columns to use in the facet.}

\item{.select}{a character vector indicating which columns to select.}
}
\value{
A ggplot object representing the time series.
}
\description{
This function creates a ggplot object for time series data.
}
\examples{
\dontrun{
# Load the 'nottem' dataset
data(nottem)

# Create a time series plot
ts_ggplot(nottem, .facet = TRUE, .hline = 50)
}

}
