% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_merge.R
\name{ts_merge}
\alias{ts_merge}
\title{ts_merge: Merge Multiple Time Series
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ts_merge(data)
}
\arguments{
\item{data}{A list of time series objects to be merged.}
}
\value{
A merged time series object.
}
\description{
Given a list of time series objects, this function merges them based on the "date" column
and returns a new time series object with all the data. If the input time series objects are not of class
"mts", then the column names will be used to label the respective series in the merged data.
}
\examples{
ts1 <- ts(rnorm(100), start = c(2010, 1), frequency = 12)
ts2 <- ts(rnorm(50), start = c(2010, 1), frequency = 12)
ts3 <- ts(rnorm(120), start = c(2010, 1), frequency = 12)
ts_merge(list(ts1, ts2, ts3))
}
