% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_select.R
\name{ts_select}
\alias{ts_select}
\title{ts_select: Select columns from a time series
r lifecycle::badge("experimental")}
\usage{
ts_select(data, .include = NULL, .exclude = NULL)
}
\arguments{
\item{data}{A time series object.}

\item{.include}{Character vector of column names to include.}

\item{.exclude}{Character vector of column names to exclude.}
}
\value{
A time series object containing the columns selected.
}
\description{
This function allows the user to select columns from a time series object
using column names, while keeping the time series format.
}
\examples{
data(ausair)
ts_select(ausair, .include = c("Q2", "Q3", "Q4"))
ts_select(ausair, .exclude = "Q1")
}
