% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_stationarity_test.R
\name{ts_stationarity_test}
\alias{ts_stationarity_test}
\title{Test Stationarity of Time Series
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ts_stationarity_test(data, .select = NULL)
}
\arguments{
\item{data}{A time series object or a data frame with multiple time series.}

\item{.select}{An optional character vector of column names of the data frame to be used.}
}
\value{
A list of ADF test results or a single ADF test result.
}
\description{
This function performs the Augmented Dickey-Fuller (ADF) test to check if a time series is stationary.
}
\examples{
ts1 <- ts(rnorm(100), start = c(2010, 1), frequency = 12)
ts2 <- ts(rnorm(100), start = c(2010, 1), frequency = 12)
data <- cbind(ts1, ts2)
ts_stationarity_test(data)
ts_stationarity_test(data, .select = c("ts1"))
ts_stationarity_test(ts1)

}
