% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_sub.R
\name{ts_sub}
\alias{ts_sub}
\title{ts_sub: Subtract two time series
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ts_sub(ts1, ts2)
}
\arguments{
\item{ts1}{A time series object.}

\item{ts2}{A time series object.}
}
\value{
A time series object representing the difference between ts1 and ts2.
The time frame starts from the end of ts2.
}
\description{
ts_sub: Subtract two time series
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
ts1 <- ts(rnorm(100), start = c(2010, 1), frequency = 12)
ts2 <- ts(rnorm(50), start = c(2010, 1), frequency = 12)
ts_sub(ts1, ts2)
}
