% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_to_df.R
\name{ts_to_df}
\alias{ts_to_df}
\title{Convert time series to data frame
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ts_to_df(ts, as_date = TRUE, .name = NULL, ...)
}
\arguments{
\item{ts}{A time series object.}

\item{as_date}{A boolean indicating whether the output should have a date column. Default is TRUE.}

\item{.name}{A character string specifying the name of the value column. If not specified, the column name will be taken from the name attribute of the time series object.}

\item{...}{Additional arguments passed to functions.}
}
\value{
A data frame with columns date and .name.
}
\description{
This function takes a time series and converts it to a data frame with a date column and a value column.
}
\examples{
\dontrun{
data(AirPassengers)
ts_to_df(AirPassengers)
ts_to_df(AirPassengers, .name = "air_passengers")
}
}
