% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_transform.R
\name{ts_transform}
\alias{ts_transform}
\title{Time series transformation}
\usage{
ts_transform(.ts, .t, .subset = NULL)
}
\arguments{
\item{.ts}{A time series object.}

\item{.t}{A transformation function to be applied to the values of the time series.}

\item{.subset}{Optional character vector of column names of the data frame to be subsetted for transformation.}
}
\value{
A time series object with transformed values.
}
\description{
This function applies a transformation function to the values of a time series object and returns
a new time series object with the transformed values.
}
\examples{
# Create a sample time series object
ts_obj <- ts(1:10, start = c(2020, 1), frequency = 12)

# Apply a transformation function to all columns
ts_transform(ts_obj, log)

# Apply a transformation function to a subset of columns
ts_transform(ts_obj, log, .subset = c("Jan", "Feb"))

}
