% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_zero.R
\name{ts_zero}
\alias{ts_zero}
\title{Create a time series object filled with zeros}
\usage{
ts_zero(x, resize = NULL)
}
\arguments{
\item{x}{A time series object.}

\item{resize}{An optional time series object to be used for resizing the output time series object.}
}
\value{
A time series object of the same length and frequency as 'x', filled with zeros.
}
\description{
This function creates a time series object filled with zeros of the same length and frequency as the input
time series object. If a 'resize' object is provided, the resulting time series object will be resized to
the length and frequency of the 'resize' object.
}
\examples{
ts_zero(AirPassengers)
}
