% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_assign.R
\name{\%<...\%}
\alias{\%<...\%}
\title{Unpack and assign values to specified names in an environment
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
names \%<...\% values
}
\arguments{
\item{names}{A list or character vector specifying the names of variables to be assigned.}

\item{values}{A list or vector containing the values to be assigned to the names.}
}
\value{
NULL. The function performs assignments in the specified environment.
}
\description{
Unpack and assign values to specified names in an environment
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
c("x", "y") \%<...\% list(1, 2)
}
}
\seealso{
Additional arguments that can be passed to names include
\itemize{
\item{.envir}{The environment where the variables will be assigned. Defaults to the global environment.}
\item{.warn}{Logical flag indicating whether to show a warning message. Defaults to TRUE.}
}
}
