% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_assign.R
\name{\%...>\%}
\alias{\%...>\%}
\title{Assign values to specified names in an environment and unpack the values (reverse of \verb{\%<...\%})}
\usage{
values \%...>\% names
}
\arguments{
\item{values}{A list or vector containing the values to be assigned to the names.}

\item{names}{A list or character vector specifying the names of variables to be assigned.}
}
\value{
NULL. The function performs assignments in the specified environment.
}
\description{
Assign values to specified names in an environment and unpack the values (reverse of \verb{\%<...\%})
}
\examples{
\dontrun{
list(1, 2) \%...>\% c("x", "y")
}
}
