% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_train_test_split.R
\name{valid_split}
\alias{valid_split}
\title{Get valiation set from a dataset split (if exists)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
valid_split(tts)
}
\arguments{
\item{tts}{\link{list}: dataset split}
}
\value{
\link{data.frame}: validation set
}
\description{
Get valiation set from a dataset split (if exists)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
warpbreaks
tts <- Dmisc::train_test_split(tbl, test_size = 0.2)
valid <- valid_split(tts)
}
}
