% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_to_date.R
\name{vars_to_date}
\alias{vars_to_date}
\title{Multiple variables to unique date variable}
\usage{
vars_to_date(
  tbl,
  year = NULL,
  trim = deprecated(),
  month = NULL,
  month_type = deprecated(),
  day = NULL
)
}
\arguments{
\item{tbl}{data.frame or tbl connection}

\item{year}{year variable position or name}

\item{trim}{trimester variable position or name}

\item{month}{month variable position or name}

\item{month_type}{character the char or num type of month variable}

\item{day}{day variable position or name}
}
\value{
tbl a new data.frame with the compute variable
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
tbl <- data.frame(
year = rep("2021", 12),
month = 1:12,
day = sample(1:3, 12, TRUE),
value = sample(100:1000, 12, TRUE)
)

tbl

vars_to_date(tbl)
}
