% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtdAvgVol.R
\name{wtdAvgVol}
\alias{wtdAvgVol}
\title{Compute the weighted average volatility of assets in a portfolio.}
\usage{
wtdAvgVol(w, R, equal.wt = TRUE, scale = 1)
}
\arguments{
\item{w}{a weight vector equal in length to the numnber of columns of R.}

\item{R}{an xts object of returns.}

\item{equal.wt}{a flag indicating whether or not to apply equal weights.
Defualt is TRUE.}

\item{scale}{the annualization factor for the data. If supplied, this will
scale the return volatilities to annualized volatilities. Default is 1.}
}
\value{
a numeric value equal to the weighted average volatility of the
    assets.
}
\description{
Given a weight vector, \code{w}, and an xts object of returns, \code{R} this
will return the weighted average volatility assuming all the assets are
perfectly uncorrelated. (ie. all pairwise covariances are zero)
If \code{equal.wt=TRUE}, then \code{w} is ignored. This is the default.
Otherwise, \code{w} is used for the weight vector.
}
\details{
This function is useful for calculations like the Diversification Ratio
where the weighted average volatility is in the numerator and the fully
correlated volatility of the portfolio is in the denominator.
}
\examples{
dts <- seq(Sys.Date()-199, Sys.Date(), 1)
returns <- matrix(rnorm(1000),ncol = 5) / 100
ret.xts <- xts(returns, dts)
wtdAvgVol(R = ret.xts)

}
\seealso{
portfolioVol, divRatio
}

