% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xtsAlign.R
\name{xtsAlign}
\alias{xtsAlign}
\title{Align one xts object to another without merging them.}
\usage{
xtsAlign(target, match.to, join = "inner")
}
\arguments{
\item{target}{the xts object you would like to modify.}

\item{match.to}{the xts object you would like to emulate or align to.}

\item{join}{the type of database join to use in matching. Default is "inner". Other
  methods could lead to strange results.}
}
\value{
an xts object whose date index is now aligned with \code{match.to}.
}
\description{
This function aligns two xts objects so that they will have the same
date indexing. The default join method is "inner." See the documentation
for merge.xts for different join methods, although this operation probably
only makes sense using an "inner" join.
}
\examples{
dts.long <- seq(Sys.Date()-49, Sys.Date(), 1)
ret.long <- matrix(rnorm(50),ncol = 1) / 100
long.xts <- xts(ret.long, dts.long)

dts.short <- seq(Sys.Date()-9, Sys.Date(), 1)
ret.short <- matrix(rnorm(10),ncol = 1) / 100
short.xts <- xts(ret.short, dts.short)

xtsAlign(target = long.xts, match.to = short.xts)
}

