% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xtsApply.R
\name{xtsApply}
\alias{xtsApply}
\title{An apply method for xts objects.}
\usage{
xtsApply(x, cFUN, margin = 2, ...)
}
\arguments{
\item{x}{an xts object over which you want to apply some function.}

\item{cFUN}{the function you wish to apply.}

\item{margin}{controls whether you want column-wise apply (margin = 2) or
  row-wise apply (margin = 1). Defaults to margin = 2.}
}
\value{
an xts object where cFUN has been applied to the columns or rows.
}
\description{
This function implements an apply method for xts objects. I works in exactly
the same was as apply does, but makes sure you get back an xts object. In
some cases the standard apply mehtods return strange resutls when applied to
xts objects. This function handles the operations correctly and will do
column-wise apply when \code{"margin"} is set to 2, and row-wise when
\code{"margin"} is set to 1. \code{"cFUN"} fills the roll of the the standard
FUN in apply methods.
}
\examples{
dts <- seq(Sys.Date()-4, Sys.Date(), 1)
returns <- matrix(rnorm(10),ncol = 2) / 100
ret.xts <- xts(returns, dts)
xtsApply(ret.xts, cFUN = sum, margin = 1)

Note that apply(ret.xts,2,sum) will return a vector, NOT and xts object!
}

