% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtsToDataFrame.R
\name{xtsToDataFrame}
\alias{xtsToDataFrame}
\title{Convert and xts object to a data frame, suitable for use in ggplot2.}
\usage{
xtsToDataFrame(xts.obj)
}
\arguments{
\item{xts.obj}{an xts object.}
}
\value{
a dataframe suitable for plotting in ggplot2. Date will be first
  column.
}
\description{
A function to convert an xts object into a data frame that can be plotted
usinfg ggplot2. The date column will be the first column of the new data
frame and will have the name 'Date'.  This function uses a recursive
implementation for xts objects with many columns.
}
\examples{
dts <- seq(Sys.Date()-4, Sys.Date(), 1)
returns <- matrix(rnorm(10),ncol = 2) / 100
ret.xts <- xts(returns, dts)
names(ret.xts) <- c("Asset_1","Asset_2")
xtsToDataFrame(ret.xts)
}

