% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_button.R
\name{big_button}
\alias{big_button}
\title{Insert a HTML big button}
\usage{
big_button(
  text = "Get started",
  href = "./articles/Dmisc.html",
  bg_color = "#00a65a"
)
}
\arguments{
\item{text}{\link{character}: text to display in the button}

\item{href}{\link{character}: HTML a tag href}

\item{bg_color}{\link{character}: HEX code for the button color}
}
\value{
a HTML element button
}
\description{
Insert a HTML big button
}
\examples{
\dontrun{
  big_button(text = "Get started", href = "./articles/Dmisc.html", bg_color = "#00a65a")
}
}
