% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_button.R
\name{big_button}
\alias{big_button}
\title{Insert a SVG/HTML big button
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
big_button(
  text,
  href,
  suffix = "",
  width = "50\%",
  height = 30,
  font_size = 18,
  font_family = NULL,
  bg_color = "#00a65a"
)
}
\arguments{
\item{text}{\link{character}: text to display in the button}

\item{href}{\link{character}: HTML a tag href}

\item{suffix}{\link{character}: suffix to display after the text}

\item{width}{width of the button}

\item{height}{height of the button}

\item{font_size}{font size of the text}

\item{font_family}{font family of the text}

\item{bg_color}{background color of the button}
}
\value{
a SVG/HTML element button
}
\description{
Insert a SVG/HTML big button
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
big_button(text = "Get started", href = "./articles/Dmisc.html")
}
}
