% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextab_functions.R
\name{bioavr_group_tab}
\alias{bioavr_group_tab}
\title{Create a grouped flextable object}
\usage{
bioavr_group_tab(flextab, header, footer)
}
\arguments{
\item{flextab}{data. Dataframe.}

\item{header}{header text. String.}

\item{footer}{footer text. String.}
}
\value{
A flextable object.
}
\description{
Usually used to create crude and age-, and sex-adjusted incidence rate tables.
Useful for exporting to word using the \code{flextable::save_as_docx()} function.
}
\examples{
bind_rows(list("Crude" = rates_table,
"Age adjusted" = rates_table_adjust), .id = "id") \%>\%
 as_grouped_data(groups = c("id")) \%>\%
 as_flextable(hide_grouplabel = T) \%>\%
 bioavr_group_tab(header = header, footer = footer)


}
