% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut3.R
\name{cut3}
\alias{cut3}
\title{Convert numeric to factor weighted and by group
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
cut3(tbl, var_name, breaks, groups = NULL, bf_args = list(), .inf = FALSE, ...)
}
\arguments{
\item{tbl}{\link{data.frame}: Database connection or data.frame}

\item{var_name}{\link{character}: variable name}

\item{breaks}{\link{numeric}: break points. See \code{base::\link[base:cut]{cut}}}

\item{groups}{\link{character}: name of a groups variable}

\item{bf_args}{\link{list}: arguments to be passed to break function}

\item{.inf}{\link{logical}: indicates if the breaks need to be extended by
-Inf and Inf}

\item{...}{argument passed to \code{base::\link[base:cut]{cut}}}
}
\value{
same as \code{tbl} input with \code{var_name} converted to factor
}
\description{
Convert numeric to factor weighted and by group
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
datos <- data.frame(edad = seq(1:100))
dplyr::count(cut3(datos, "edad", 5), edad)
}
}
\seealso{
\code{vignette("cut3", package = "Dmisc")}
\code{base::\link[base:cut]{cut}}
}
