% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut3.R
\name{cut3}
\alias{cut3}
\title{Convert numeric to factor weighted and by group}
\usage{
cut3(tbl, var_name, breaks, groups = NULL, ...)
}
\arguments{
\item{tbl}{\link{data.frame}: Database connection or data.frame}

\item{var_name}{\link{character}: variable name}

\item{breaks}{\link{numeric}: break points. See \code{base::\link[base:cut]{cut}}}

\item{groups}{\link{character}: name of a numeric variable of groups}

\item{...}{argument passed to cut and/or breaks functions}
}
\value{
same as \code{tbl} input with \code{var_name} converted to factor
}
\description{
Convert numeric to factor weighted and by group
}
\examples{
datos <- data.frame(edad = seq(1:100))
cut3(datos, "edad", 5)
}
\seealso{
\code{vignette("cut3", package = "Dmisc")}
\code{base::\link[base:cut]{cut}}
}
