% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Create a table of summaries for an entery dataframe}
\usage{
describe(data, digits = 4, t = TRUE, flextable = FALSE, ft_args = list(), ...)
}
\arguments{
\item{data}{\link{data.frame}: a dataframe with the data}

\item{digits}{\link{numeric}: number of digits for the numbers}

\item{t}{\link{logical}: indicates if the resulting table is transposed}

\item{flextable}{\link{logical}: indicates if the table is converted to a simple flextable.
See \code{flextable::\link[flextable:flextable]{flextable}}}

\item{ft_args}{\link{list}: other arguments passed to \code{flextable::\link[flextable:flextable]{flextable}} function}

\item{...}{other arguments passed to summary function}
}
\value{
a table containing the summary of every variable in the provided dataframe
}
\description{
\lifecycle{stable}
}
\examples{
\dontrun{
describe(cars, flextable = T)
describe(cars, flextable = T, ft_args = list(cwidth = 1))
}
}
