% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Create a summary table for a data frame
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
describe(data, digits = 4, t = TRUE, flextable = FALSE, ft_args = list(), ...)
}
\arguments{
\item{data}{A data frame containing the data to be summarized}

\item{digits}{The number of digits to display in the summary table (default is 4)}

\item{t}{If TRUE, the table will be transposed (default is TRUE)}

\item{flextable}{If TRUE, the table will be converted to a simple flextable (default is FALSE)}

\item{ft_args}{A list of additional arguments to pass to the \code{flextable} function if \code{flextable = TRUE} (default is an empty list)}

\item{...}{Other arguments to pass to the \code{summary} function}
}
\value{
A table containing summary statistics for each variable in the data frame
}
\description{
This function generates a summary table for a data frame, containing summary statistics for each variable in the data frame.
}
\examples{
\dontrun{
describe(cars, flextable = TRUE)
describe(cars, flextable = TRUE, ft_args = list(cwidth = 1))
}
}
