% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArgs.R
\name{getArgs}
\alias{getArgs}
\title{Get function arguments}
\usage{
getArgs()
}
\description{
Gets the function arguments nicely.  Does not return arguments
that are set by defaults
}
\examples{
test <- function(x=1, y=2, z=3){getArgs()}
}
\references{
\url{http://stackoverflow.com/a/17257053/1003565}
}
\author{
Hong Ooi (from StackOverflow)
}
