% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{A wrapper for getSymbols that can handle lists of tickers.}
\usage{
getData(tickers, datasrc, quiet = FALSE, ...)
}
\arguments{
\item{tickers}{a character vector of symbols to be passed to getSymbols.}

\item{datasrc}{the data source for the call. Usually, "yahoo" or "google" or "FRED".}

\item{quiet}{controls verboseness of getSymbols call.}
}
\description{
This is just a simple wrapper for getSymbols that can take either a
single ticker vector or a list of ticker vectors for, say, multiple
asset classes.
}

