% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_assign.R
\name{\%...>\%}
\alias{\%...>\%}
\title{Wrapper around \verb{\%<...\%} to unpack and assign values to specified names in an environment
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
values \%...>\% names
}
\arguments{
\item{values}{A list or vector containing the values to be assigned to the names.}

\item{names}{A character vector specifying the names of variables to be assigned.}

\item{.envir}{The environment where the variables will be assigned. Defaults to the global environment.}

\item{.warn}{Logical flag indicating whether to show a warning message. Defaults to TRUE.}
}
\value{
NULL. The function performs assignments in the specified environment.
}
\description{
Wrapper around \verb{\%<...\%} to unpack and assign values to specified names in an environment
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
values <- list(1, 2)
names <- c("x", "y")
values \%...>\% names
}
