% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holtWinters.R
\name{holtWinters}
\alias{holtWinters}
\title{Holt winters smoothing no seasonality}
\usage{
holtWinters(y, alpha = 0.5, beta = 0.5)
}
\arguments{
\item{y}{Vector. The data to smooth.}

\item{alpha}{Numeric.  A smoothing parameter}

\item{beta}{Numeric. A smoothing parameter}
}
\description{
Holt winters smoothing with no seasonality
}
\details{
I also don't trust this code anymore. But I don't use it so whatever.
}
