% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertHTML.R
\name{insertHTML}
\alias{insertHTML}
\title{Insert HTML file}
\usage{
insertHTML(path, name)
}
\arguments{
\item{path}{character the path where the HTML file exists. May end with /}

\item{name}{character the name of HTML file to insert}
}
\value{
character an HTML character string
}
\description{
\lifecycle{experimental}
}
\details{
This function was created to insert MatLab published HTML into
RMarkdown documents. Important considerations:
\itemize{
\item{
Only the body of the document if returned.
}
\item{
The images are converted to base64 strings.
}
\item{
A caption with de file name is added to each image.
}
\item{
The output is wrapped by a \code{<pre></pre>} tag.
}
}
}
