% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeNoise.R
\name{makeNoise}
\alias{makeNoise}
\title{Create a set of returns in xts format that follow a specified distribution.}
\usage{
makeNoise(n, dates, num_assets, dist, ...)
}
\arguments{
\item{n}{the number of returns you would like returned.}

\item{dates}{the dates you would like to use for the returns. If NULL, the
function will use \code{SysDate()} and go back in time \code{n} days.}

\item{num_assets}{the number of columns of length \code{n} to create.}

\item{dist}{one of "rnorm", "runif", or "dt".}

\item{...}{other pass-through parameters to the random number generators.}
}
\value{
a set of returns in xts format following the desired distribution.
}
\description{
This function returns a set of returns that follow the distribution passed
to makeNoise by the user. It is a convenience function when you want a set
of random returns that follow a certain distribution. The function is a
wrapper for rnorm etc. that handles all the time series wrangling.
}
\examples{
# Make 5 normally distributed assets...
makeNoise(10, dates = NULL, num_assets = 5, dist = "rnorm", mean = .05/252,
sd = .10/sqrt(252))
}

