% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_bar.R
\name{progress_bar}
\alias{progress_bar}
\title{Html progress bar}
\usage{
progress_bar(complete, total)
}
\arguments{
\item{complete}{\link{numeric}: number of items completed}

\item{total}{\link{numeric}: number of total items}
}
\value{
a html object representing a progress bar
}
\description{
Add a html 4 colors progress bar.
}
\examples{
\dontrun{
  progress_bar(77, 102)
}
}
