% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sys_Date_locale.R
\name{Sys_Date_locale}
\alias{Sys_Date_locale}
\title{System date with specific format and locale}
\usage{
Sys_Date_locale(locale = "es_DO.utf8", format = "\%B \%d, \%Y")
}
\arguments{
\item{locale}{a currently installed locale string}

\item{format}{a valid format for dates or NULL}
}
\value{
a date with specified locale and format
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
You need to have installed the specified locale.
}
\examples{
\dontrun{
 Sys_Date_locale()
}
}
