% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{Split dataset into train and test sets
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
train_test_split(tbl, test_size = 0.2, valid_split = FALSE, ts = FALSE)
}
\arguments{
\item{tbl}{\link{data.frame}: dataset to split}

\item{test_size}{\link{numeric}: proportion of the dataset to use as test set}

\item{valid_split}{\link{logical}: if TRUE, the test set will be splitted into validation and test set}

\item{ts}{\link{logical}: if TRUE, indicates that the dataset is time series}
}
\value{
\link{list}: a list with the data and the indices for the splits
}
\description{
Split dataset into train and test sets
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
 warpbreaks
 tts <- train_test_split(tbl, test_size = 0.2)
 train <- Dmisc::train_split(tts)
 test <- Dmisc::test_split(tts)
}
}
