% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_renv.R
\name{use_renv}
\alias{use_renv}
\title{Initialize renv}
\usage{
use_renv(envs_path = "~/.renv/library", ...)
}
\arguments{
\item{envs_path}{path where you want to save the libraries.
The default value is relative to R's home directory.}

\item{...}{other arguments passed to \code{renv::\link[renv:init]{init}}}
}
\value{
\link{NULL}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This function does the same as \code{renv::\link[renv:init]{init}} but forces renv
to use an external (customizable) library location.
}
\examples{
\dontrun{
 use_renv
}
}
