% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_to_date.R
\name{vars_to_date}
\alias{vars_to_date}
\title{Multiple variables to unique date variable
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
vars_to_date(
  tbl,
  year = NULL,
  quarter = NULL,
  month = NULL,
  day = NULL,
  date = NULL,
  drop_vars = TRUE,
  clean_names = FALSE,
  date_format = "\%d-\%m-\%y",
  origin = "1900-01-01",
  .round = c("end", "middle", "start")
)
}
\arguments{
\item{tbl}{data.frame or tbl connection}

\item{year}{year variable position or name}

\item{quarter}{quarter variable position or name}

\item{month}{month variable position or name}

\item{day}{day variable position or name}

\item{date}{a variable name or position containing a like date format}

\item{drop_vars}{indicates if variables should be dropped}

\item{clean_names}{indicates if all variable names should be cleaned}

\item{date_format}{actual date format of variable in \code{date} argument}

\item{origin}{base date for variable convertion to date}
}
\value{
tbl a new data.frame with the compute variable
}
\description{
Multiple variables to unique date variable
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
tbl <- data.frame(
  year = rep("2021", 12),
  month = month.name,
  day = sample(1:3, 12, TRUE),
  value = sample(100:1000, 12, TRUE)
)

tbl

vars_to_date(tbl, year = 1, month = 2, day = 3)
}
